+++ Python/Product/TestAdapter.Executor/Pytest/PytestTestDiscoverer.cs
@@ -13,21 +13,6 @@
 //
 // See the Apache Version 2.0 License for specific language governing
 // permissions and limitations under the License.
-// Python Tools for Visual Studio
-// Copyright(c) Microsoft Corporation
-// All rights reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the License); you may not use
-// this file except in compliance with the License. You may obtain a copy of the
-// License at http://www.apache.org/licenses/LICENSE-2.0
-//
-// THIS CODE IS PROVIDED ON AN  *AS IS* BASIS, WITHOUT WARRANTIES OR CONDITIONS
-// OF ANY KIND, EITHER EXPRESS OR IMPLIED, INCLUDING WITHOUT LIMITATION ANY
-// IMPLIED WARRANTIES OR CONDITIONS OF TITLE, FITNESS FOR A PARTICULAR PURPOSE,
-// MERCHANTABILITY OR NON-INFRINGEMENT.
-//
-// See the Apache Version 2.0 License for specific language governing
-// permissions and limitations under the License.
 
 using System;
 using System.Collections.Generic;
@@ -42,13 +27,13 @@ using Microsoft.VisualStudio.TestPlatform.ObjectModel.Adapter;
 using Microsoft.VisualStudio.TestPlatform.ObjectModel.Logging;
 using Newtonsoft.Json;
 
-namespace Microsoft.PythonTools.TestAdapter.UnitTest {
-    internal class TestDiscovererUnitTest : IPythonTestDiscoverer {
+namespace Microsoft.PythonTools.TestAdapter.Pytest {
+    internal class PytestTestDiscoverer : IPythonTestDiscoverer {
         private readonly PythonProjectSettings _settings;
         private IMessageLogger _logger;
         private static readonly string DiscoveryAdapterPath = PythonToolsInstallPath.GetFile("PythonFiles\\testing_tools\\run_adapter.py");
 
-        public TestDiscovererUnitTest(PythonProjectSettings settings) {
+        public PytestTestDiscoverer(PythonProjectSettings settings) {
             _settings = settings;
         }
 
@@ -65,14 +50,14 @@ namespace Microsoft.PythonTools.TestAdapter.UnitTest {
                 throw new ArgumentNullException(nameof(discoverySink));
             }
 
-            _logger = logger ?? throw new ArgumentNullException(nameof(logger)); 
+            _logger = logger ?? throw new ArgumentNullException(nameof(logger));
 
             var workspaceText = _settings.IsWorkspace ? Strings.WorkspaceText : Strings.ProjectText;
-            LogInfo(Strings.PythonTestDiscovererStartedMessage.FormatUI(PythonConstants.UnitTestText, _settings.ProjectName, workspaceText, _settings.DiscoveryWaitTimeInSeconds));
+            LogInfo(Strings.PythonTestDiscovererStartedMessage.FormatUI(PythonConstants.PytestText, _settings.ProjectName, workspaceText, _settings.DiscoveryWaitTimeInSeconds));
 
             var env = InitializeEnvironment(sources, _settings);
             var outputFilePath = Path.GetTempFileName();
-            var arguments = GetArguments(sources, outputFilePath);
+            var arguments = GetArguments(sources, _settings, outputFilePath);
 
             LogInfo("cd " + _settings.WorkingDirectory);
             LogInfo("set " + _settings.PathEnv + "=" + env[_settings.PathEnv]);
@@ -101,12 +86,12 @@ namespace Microsoft.PythonTools.TestAdapter.UnitTest {
             }
 
             string json = File.ReadAllText(outputFilePath);
-            if (String.IsNullOrEmpty(json)) {
+            if (string.IsNullOrEmpty(json)) {
                 return;
             }
 
             try {
-                var results = JsonConvert.DeserializeObject<List<UnitTestDiscoveryResults>>(json);
+                var results = JsonConvert.DeserializeObject<List<PytestDiscoveryResults>>(json);
                 var testcases = ParseDiscoveryResults(results);
 
                 foreach (var tc in testcases) {
@@ -124,20 +109,21 @@ namespace Microsoft.PythonTools.TestAdapter.UnitTest {
             }
         }
 
-        internal IEnumerable<TestCase> ParseDiscoveryResults(IList<UnitTestDiscoveryResults> results) {
+        internal IEnumerable<TestCase> ParseDiscoveryResults(IList<PytestDiscoveryResults> results) {
             if (results is null) {
                 throw new ArgumentNullException(nameof(results));
             }
 
-            return results
+            var testcases = results
                 .Where(r => r.Tests != null)
                 .SelectMany(r => r.Tests.Select(test => TryCreateVsTestCase(test)))
                 .Where(tc => tc != null);
+
+            return testcases;
         }
 
-        private TestCase TryCreateVsTestCase(UnitTestTestCase test) {
+        private TestCase TryCreateVsTestCase(PytestTest test) {
             try {
-                // Note: Test Explorer will show a key not found exception if we use a source path that doesn't match a test container's source.
                 TestCase tc = test.ToVsTestCase(_settings.ProjectHome);
                 return tc;
             } catch (Exception ex) {
@@ -146,18 +132,17 @@ namespace Microsoft.PythonTools.TestAdapter.UnitTest {
             return null;
         }
 
-        public string[] GetArguments(IEnumerable<string> sources, string outputfilename) {
+        public string[] GetArguments(IEnumerable<string> sources, PythonProjectSettings projSettings, string outputfilename) {
             var arguments = new List<string>();
             arguments.Add(DiscoveryAdapterPath);
             arguments.Add("discover");
-            arguments.Add("unittest");
+            arguments.Add("pytest");
             arguments.Add("--output-file");
             arguments.Add(outputfilename);
-            //Note unittest specific options go after this separator
+            //Note pytest specific arguments go after this separator
             arguments.Add("--");
-            arguments.Add(_settings.UnitTestRootDir);
-            arguments.Add(_settings.UnitTestPattern);
-
+            arguments.Add("--cache-clear");
+            arguments.Add(String.Format("--rootdir={0}", projSettings.ProjectHome));
             return arguments.ToArray();
         }
 
@@ -175,12 +160,12 @@ namespace Microsoft.PythonTools.TestAdapter.UnitTest {
             }
 
             env["PYTHONUNBUFFERED"] = "1";
-
             return env;
         }
 
         private string GetSearchPaths(IEnumerable<string> sources, PythonProjectSettings settings) {
             var paths = settings.SearchPath;
+
             paths.Insert(0, settings.WorkingDirectory);
 
             string searchPaths = string.Join(
@@ -192,19 +177,18 @@ namespace Microsoft.PythonTools.TestAdapter.UnitTest {
 
         [Conditional("DEBUG")]
         private void DebugInfo(string message) {
-            _logger?.SendMessage(TestMessageLevel.Informational, message);
+            _logger?.SendMessage(TestMessageLevel.Informational, message ?? String.Empty);
         }
 
         private void LogInfo(string message) {
-            _logger?.SendMessage(TestMessageLevel.Informational, message);
+            _logger?.SendMessage(TestMessageLevel.Informational, message ?? String.Empty);
         }
 
         private void Error(string message) {
-            _logger?.SendMessage(TestMessageLevel.Error, message);
+            _logger?.SendMessage(TestMessageLevel.Error, message ?? String.Empty);
         }
-
         private void Warn(string message) {
-            _logger?.SendMessage(TestMessageLevel.Warning, message);
+            _logger?.SendMessage(TestMessageLevel.Warning, message ?? String.Empty);
         }
     }
 }
