+++ Python/Product/TestAdapter.Executor/TestAdapter.Executor.csproj
@@ -93,8 +93,8 @@
     <Compile Include="Common\Config\RunSettingsUtil.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Pytest\PytestExtensions.cs" />
-    <Compile Include="Pytest\TestDiscovererPytest.cs" />
-    <Compile Include="Pytest\TestExecutorPytest.cs" />
+    <Compile Include="Pytest\PytestTestDiscoverer.cs" />
+    <Compile Include="Pytest\PytestTestExecutor.cs" />
     <Compile Include="Common\Services\DiscovererFactory.cs" />
     <Compile Include="Common\Services\ExecutorService.cs" />
     <Compile Include="Pytest\JunitXmlTestResultParser.cs" />
@@ -106,8 +106,8 @@
     <Compile Include="Common\PythonTestDiscoverer.cs" />
     <Compile Include="Common\TestProtocol.cs" />
     <Compile Include="Common\TestReader.cs" />
-    <Compile Include="UnitTest\TestExecutorUnitTest.cs" />
-    <Compile Include="UnitTest\TestDiscovererUnitTest.cs" />
+    <Compile Include="UnitTest\UnitTestTestExecutor.cs" />
+    <Compile Include="UnitTest\UnitTestTestDiscoverer.cs" />
     <Compile Include="UnitTest\UnitTestDiscoveryResults.cs" />
     <Compile Include="UnitTest\UnitTestExtensions.cs" />
     <Compile Include="Common\Utils\TestUtils.cs" />
