+++ Python/Product/TestAdapter.Executor/UnitTest/UnitTestTestDiscoverer.cs
@@ -13,6 +13,21 @@
 //
 // See the Apache Version 2.0 License for specific language governing
 // permissions and limitations under the License.
+// Python Tools for Visual Studio
+// Copyright(c) Microsoft Corporation
+// All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the License); you may not use
+// this file except in compliance with the License. You may obtain a copy of the
+// License at http://www.apache.org/licenses/LICENSE-2.0
+//
+// THIS CODE IS PROVIDED ON AN  *AS IS* BASIS, WITHOUT WARRANTIES OR CONDITIONS
+// OF ANY KIND, EITHER EXPRESS OR IMPLIED, INCLUDING WITHOUT LIMITATION ANY
+// IMPLIED WARRANTIES OR CONDITIONS OF TITLE, FITNESS FOR A PARTICULAR PURPOSE,
+// MERCHANTABILITY OR NON-INFRINGEMENT.
+//
+// See the Apache Version 2.0 License for specific language governing
+// permissions and limitations under the License.
 
 using System;
 using System.Collections.Generic;
@@ -27,13 +42,13 @@ using Microsoft.VisualStudio.TestPlatform.ObjectModel.Adapter;
 using Microsoft.VisualStudio.TestPlatform.ObjectModel.Logging;
 using Newtonsoft.Json;
 
-namespace Microsoft.PythonTools.TestAdapter.Pytest {
-    internal class TestDiscovererPytest : IPythonTestDiscoverer {
+namespace Microsoft.PythonTools.TestAdapter.UnitTest {
+    internal class UnitTestTestDiscoverer : IPythonTestDiscoverer {
         private readonly PythonProjectSettings _settings;
         private IMessageLogger _logger;
         private static readonly string DiscoveryAdapterPath = PythonToolsInstallPath.GetFile("PythonFiles\\testing_tools\\run_adapter.py");
 
-        public TestDiscovererPytest(PythonProjectSettings settings) {
+        public UnitTestTestDiscoverer(PythonProjectSettings settings) {
             _settings = settings;
         }
 
@@ -50,14 +65,14 @@ namespace Microsoft.PythonTools.TestAdapter.Pytest {
                 throw new ArgumentNullException(nameof(discoverySink));
             }
 
-            _logger = logger ?? throw new ArgumentNullException(nameof(logger));
+            _logger = logger ?? throw new ArgumentNullException(nameof(logger)); 
 
             var workspaceText = _settings.IsWorkspace ? Strings.WorkspaceText : Strings.ProjectText;
-            LogInfo(Strings.PythonTestDiscovererStartedMessage.FormatUI(PythonConstants.PytestText, _settings.ProjectName, workspaceText, _settings.DiscoveryWaitTimeInSeconds));
+            LogInfo(Strings.PythonTestDiscovererStartedMessage.FormatUI(PythonConstants.UnitTestText, _settings.ProjectName, workspaceText, _settings.DiscoveryWaitTimeInSeconds));
 
             var env = InitializeEnvironment(sources, _settings);
             var outputFilePath = Path.GetTempFileName();
-            var arguments = GetArguments(sources, _settings, outputFilePath);
+            var arguments = GetArguments(sources, outputFilePath);
 
             LogInfo("cd " + _settings.WorkingDirectory);
             LogInfo("set " + _settings.PathEnv + "=" + env[_settings.PathEnv]);
@@ -86,12 +101,12 @@ namespace Microsoft.PythonTools.TestAdapter.Pytest {
             }
 
             string json = File.ReadAllText(outputFilePath);
-            if (string.IsNullOrEmpty(json)) {
+            if (String.IsNullOrEmpty(json)) {
                 return;
             }
 
             try {
-                var results = JsonConvert.DeserializeObject<List<PytestDiscoveryResults>>(json);
+                var results = JsonConvert.DeserializeObject<List<UnitTestDiscoveryResults>>(json);
                 var testcases = ParseDiscoveryResults(results);
 
                 foreach (var tc in testcases) {
@@ -109,21 +124,20 @@ namespace Microsoft.PythonTools.TestAdapter.Pytest {
             }
         }
 
-        internal IEnumerable<TestCase> ParseDiscoveryResults(IList<PytestDiscoveryResults> results) {
+        internal IEnumerable<TestCase> ParseDiscoveryResults(IList<UnitTestDiscoveryResults> results) {
             if (results is null) {
                 throw new ArgumentNullException(nameof(results));
             }
 
-            var testcases = results
+            return results
                 .Where(r => r.Tests != null)
                 .SelectMany(r => r.Tests.Select(test => TryCreateVsTestCase(test)))
                 .Where(tc => tc != null);
-
-            return testcases;
         }
 
-        private TestCase TryCreateVsTestCase(PytestTest test) {
+        private TestCase TryCreateVsTestCase(UnitTestTestCase test) {
             try {
+                // Note: Test Explorer will show a key not found exception if we use a source path that doesn't match a test container's source.
                 TestCase tc = test.ToVsTestCase(_settings.ProjectHome);
                 return tc;
             } catch (Exception ex) {
@@ -132,17 +146,18 @@ namespace Microsoft.PythonTools.TestAdapter.Pytest {
             return null;
         }
 
-        public string[] GetArguments(IEnumerable<string> sources, PythonProjectSettings projSettings, string outputfilename) {
+        public string[] GetArguments(IEnumerable<string> sources, string outputfilename) {
             var arguments = new List<string>();
             arguments.Add(DiscoveryAdapterPath);
             arguments.Add("discover");
-            arguments.Add("pytest");
+            arguments.Add("unittest");
             arguments.Add("--output-file");
             arguments.Add(outputfilename);
-            //Note pytest specific arguments go after this separator
+            //Note unittest specific options go after this separator
             arguments.Add("--");
-            arguments.Add("--cache-clear");
-            arguments.Add(String.Format("--rootdir={0}", projSettings.ProjectHome));
+            arguments.Add(_settings.UnitTestRootDir);
+            arguments.Add(_settings.UnitTestPattern);
+
             return arguments.ToArray();
         }
 
@@ -160,12 +175,12 @@ namespace Microsoft.PythonTools.TestAdapter.Pytest {
             }
 
             env["PYTHONUNBUFFERED"] = "1";
+
             return env;
         }
 
         private string GetSearchPaths(IEnumerable<string> sources, PythonProjectSettings settings) {
             var paths = settings.SearchPath;
-
             paths.Insert(0, settings.WorkingDirectory);
 
             string searchPaths = string.Join(
@@ -177,18 +192,19 @@ namespace Microsoft.PythonTools.TestAdapter.Pytest {
 
         [Conditional("DEBUG")]
         private void DebugInfo(string message) {
-            _logger?.SendMessage(TestMessageLevel.Informational, message ?? String.Empty);
+            _logger?.SendMessage(TestMessageLevel.Informational, message);
         }
 
         private void LogInfo(string message) {
-            _logger?.SendMessage(TestMessageLevel.Informational, message ?? String.Empty);
+            _logger?.SendMessage(TestMessageLevel.Informational, message);
         }
 
         private void Error(string message) {
-            _logger?.SendMessage(TestMessageLevel.Error, message ?? String.Empty);
+            _logger?.SendMessage(TestMessageLevel.Error, message);
         }
+
         private void Warn(string message) {
-            _logger?.SendMessage(TestMessageLevel.Warning, message ?? String.Empty);
+            _logger?.SendMessage(TestMessageLevel.Warning, message);
         }
     }
 }
